+++ src/Security/Security/Models/IotSecuritySolutionAnalytics/PSIoTSeverityMetrics.cs
@@ -16,10 +16,10 @@ namespace Microsoft.Azure.Commands.Security.Models.IotSecuritySolutionAnalytics
 {
     public class PSIoTSeverityMetrics
     {
-        public int? High { get; set; }
+        public long? High { get; set; }
 
-        public int? Medium { get; set; }
+        public long? Medium { get; set; }
 
-        public int? Low { get; set; }
+        public long? Low { get; set; }
     }
 }
